[memory] fix warnings, refactoring package
